src/use-cases/authenticate.ts
import { CustomerRepository } from '@/repositories/customers-repository'
import { InvalidCredentialsError } from './errors/invalid-credentials-error'
import { compare } from 'bcryptjs'
import { Customer } from '@prisma/client'

interface AuthenticateUseCaseRequest {
  email: string
  password: string
}

interface AuthenticateUseCaseResponse {
  customer: Customer
}

export class AuthenticateUseCase {
  constructor(private customersRepository: CustomerRepository) {}

  async execute({
    email,
    password,
  }: AuthenticateUseCaseRequest): Promise<AuthenticateUseCaseResponse> {
    const customer = await this.customersRepository.findByEmail(email)

    if (!customer) {
      throw new InvalidCredentialsError()
    }

    const doesPasswordMatches = await compare(password, customer.password_hash)

    if (!doesPasswordMatches) {
      throw new InvalidCredentialsError()
    }

    return {
      customer,
    }
  }
}
